python_app/requirements.txt
flask
flask_cors
flask_oauth
flask_login
pymongo
requests
python-dateutil
schedule
python-dotenv
nltk
tqdm
